feat: add Recordings and VoicemailInbox components, update translations and links in VoiceMailContent
