+++ components/history/recordings/Recordings.tsx
@@ -1,0 +1,15 @@
+// Copyright (C) 2025 Nethesis S.r.l.
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+import { FC, ComponentProps } from 'react'
+
+export interface RecordingsProps extends ComponentProps<'div'> {}
+
+export const Recordings: FC<RecordingsProps> = ({ className }): JSX.Element => {
+    return (
+    <>
+    </>
+  )
+}
+
+Recordings.displayName = 'Recordings'
