+++ components/history/voicemail inbox/VoicemailInbox.tsx
@@ -1,0 +1,15 @@
+// Copyright (C) 2025 Nethesis S.r.l.
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+import { FC, ComponentProps } from 'react'
+
+export interface VoicemailInboxProps extends ComponentProps<'div'> {}
+
+export const VoicemailInbox: FC<VoicemailInboxProps> = ({ className }): JSX.Element => {
+    return (
+    <>
+    </>
+  )
+}
+
+VoicemailInbox.displayName = 'Voicemail Inbox'
